requirements.txt
flask
langchain
langchain-community
openai    # or groq llama 3 client
amadeus
python-dotenv

